# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Igloo",
    "description": "The Igloo control plane API.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/v1/repos/{id}/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Lists a repository's tasks, oldest first unless `order=newest`, optionally only those at the\ngiven phases.",
        "operationId": "listRepoTasks",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phase",
            "in": "query",
            "description": "Only tasks at this phase; repeatable",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TaskPhase"
              }
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "`newest` lists the newest first",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ListOrder"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResource"
                  }
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Creates a task: the agent starts from the head of the repository's default branch.",
        "operationId": "createTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/tasks/{id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Gets a task.",
        "operationId": "getTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{id}/cancel": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Cancels a task and stops its sandbox. Cancelling an ended task changes nothing, except that\na task that failed collecting its commits stops the sandbox it kept for recovery.",
        "operationId": "cancelTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{id}/transcript": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "A task's transcript from a position on: what its tool said and did, turn by turn, read from\nits output with secrets masked. Poll with `after` set to the last response's `next`.",
        "operationId": "getTaskTranscript",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "path",
            "description": "The first position to return; 0 when absent.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTaskRequest": {
        "type": "object",
        "description": "Creates a task on a repository.",
        "required": [
          "goal"
        ],
        "properties": {
          "goal": {
            "type": "string",
            "description": "What the agent is asked to do; not empty."
          },
          "tool": {
            "type": [
              "string",
              "null"
            ],
            "description": "A tool of the repository's `.igloo/agents.toml`; its default tool when absent."
          }
        }
      },
...
      "FieldError": {
        "type": "object",
        "description": "One invalid field of a request.",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string",
            "description": "The field path, such as `\"limits.millicpus\"`."
          },
          "message": {
            "type": "string",
            "description": "Why it is invalid."
          }
        }
      },
...
      "ListOrder": {
        "type": "string",
        "description": "The order of a list. Lists are oldest first unless asked otherwise.",
        "enum": [
          "oldest",
          "newest"
        ]
      },
...
      "Problem": {
        "type": "object",
        "description": "An error response. `code` is stable and machine-readable; `title` and `detail` are for\nhumans.",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "The stable error code, such as `\"sandbox.not_found\"`."
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "What went wrong in this occurrence."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "description": "Every invalid field, for validation problems."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status code.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "A short summary of the problem type."
          },
          "type": {
            "type": "string",
            "description": "A URI identifying the problem type."
          }
        }
      },
...
      "TaskPhase": {
        "type": "string",
        "description": "Where a task is.",
        "enum": [
          "preparing",
          "working",
          "awaiting_review",
          "done",
          "failed",
          "cancelled"
        ]
      },
      "TaskResource": {
        "type": "object",
        "description": "A task.",
        "required": [
          "id",
          "repo",
          "goal",
          "phase",
          "turns",
          "created_at"
        ],
        "properties": {
          "change": {
            "type": [
              "string",
              "null"
            ],
            "description": "The change its commits form, once it has one."
          },
          "commit": {
            "type": [
              "string",
              "null"
            ],
            "description": "The commit it started from, once settled."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When it was created."
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why it failed."
          },
          "goal": {
            "type": "string",
            "description": "What the agent is asked to do."
          },
          "id": {
            "type": "string",
            "description": "Its id (`task_...`)."
          },
          "phase": {
            "$ref": "#/components/schemas/TaskPhase",
            "description": "Where it is."
          },
          "repo": {
            "type": "string",
            "description": "The repository."
          },
          "sandbox": {
            "type": [
              "string",
              "null"
            ],
            "description": "Its sandbox, once started."
          },
          "tool": {
            "type": [
              "string",
              "null"
            ],
            "description": "The tool, once settled."
          },
          "turns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TurnResource"
            },
            "description": "Its turns, oldest first."
          }
        }
      },
      "TranscriptEntry": {
        "type": "object",
        "description": "An entry of a task's transcript.",
        "required": [
          "position",
          "turn",
          "item"
        ],
        "properties": {
          "item": {
            "$ref": "#/components/schemas/TranscriptItem",
            "description": "What happened."
          },
          "position": {
            "type": "integer",
            "format": "int32",
            "description": "Its position in the transcript, from 0.",
            "minimum": 0
          },
          "turn": {
            "type": "integer",
            "format": "int32",
            "description": "The turn it belongs to, from 1.",
            "minimum": 0
          }
        }
      },
      "TranscriptItem": {
        "oneOf": [
          {
            "type": "object",
            "description": "Text the tool wrote to the reader.",
            "required": [
              "text",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "message"
                ]
              },
              "text": {
                "type": "string",
                "description": "The text."
              }
            }
          },
          {
            "type": "object",
            "description": "The tool called one of its tools.",
            "required": [
              "id",
              "name",
              "input",
              "kind"
            ],
            "properties": {
              "id": {
                "type": "string",
                "description": "The call's id, matching its result."
              },
              "input": {
                "type": "string",
                "description": "Its input, as JSON."
              },
              "kind": {
                "type": "string",
                "enum": [
                  "tool_call"
                ]
              },
              "name": {
                "type": "string",
                "description": "The tool called, such as `Bash`."
              }
            }
          },
          {
            "type": "object",
            "description": "What a tool call returned.",
            "required": [
              "id",
              "output",
              "is_error",
              "kind"
            ],
            "properties": {
              "id": {
                "type": "string",
                "description": "The call's id."
              },
              "is_error": {
                "type": "boolean",
                "description": "Whether the call failed."
              },
              "kind": {
                "type": "string",
                "enum": [
                  "tool_result"
                ]
              },
              "output": {
                "type": "string",
                "description": "The output."
              }
            }
          },
          {
            "type": "object",
            "description": "A line of output the tool's harness does not read further.",
            "required": [
              "text",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "output"
                ]
              },
              "text": {
                "type": "string",
                "description": "The line."
              }
            }
          },
          {
            "type": "object",
            "description": "The tool reported a failure.",
            "required": [
              "text",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "error"
                ]
              },
              "text": {
                "type": "string",
                "description": "What it said."
              }
            }
          }
        ],
        "description": "One thing a tool did or said."
      },
      "TranscriptResource": {
        "type": "object",
        "description": "A task's transcript from a position on.",
        "required": [
          "entries",
          "next",
          "idle"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranscriptEntry"
            },
            "description": "The entries, oldest first."
          },
          "idle": {
            "type": "boolean",
            "description": "Whether no turn runs, so no entries follow until another turn starts."
          },
          "next": {
            "type": "integer",
            "format": "int32",
            "description": "The position to read from next.",
            "minimum": 0
          }
        }
      },
      "TurnResource": {
        "type": "object",
        "description": "One turn of a task.",
        "required": [
          "number",
          "job",
          "prompt",
          "status"
        ],
        "properties": {
          "job": {
            "type": "string",
            "description": "The job running the tool, whose logs are its output."
          },
          "number": {
            "type": "integer",
            "format": "int32",
            "description": "Its position, from 1.",
            "minimum": 0
          },
          "prompt": {
            "type": "string",
            "description": "What the tool was asked."
          },
          "reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "How it failed."
          },
          "status": {
            "$ref": "#/components/schemas/TurnStatus",
            "description": "Where it is."
          }
        }
      },
      "TurnStatus": {
        "type": "string",
        "description": "Where a turn is.",
        "enum": [
          "running",
          "succeeded",
          "failed"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}